updated WSEndPoint to TS
